Configure sixtyfour settings

Configure sixtyfour settings

Usage

aws_configure(redacted = FALSE, redact_str = "*****", verbose = TRUE)

Arguments

redacted

(logical) Redact secrets? Default: FALSE. If TRUE, secret values are redacted (replaced with redact_str) in certain messages and output from functions. See What is Redacted below.

redact_str

(character) String to use to replace redacted values. Default: "*****"

verbose

(logical) Print verbose output? Default: TRUE. Applies only to cli::cli_alert_info(), cli::cli_alert_warning(), and cli::cli_alert_success() functions that are used throughout this package. There's still a few places where verbose may not be respected.

What is Redacted

What's redacted is currently hard-coded in the package. There's only certain functions and certain elements in the output of those functions that are redacted. The following is what's redacted with aws_configure(redacted = TRUE) or with_redacted():

aws_whoami(): AWS Account ID via account_id()

six_user_creds(): Access Key ID

groups functions:

functions: aws_groups(), aws_group(), aws_group_create()

attribute: Arn (includes AWS Account ID)

roles functions:

functions: aws_roles(), aws_role(), aws_role_create()

attribute: Arn (includes AWS Account ID)

user functions:

functions: aws_users(), aws_user(), aws_user_create(), aws_user_add_to_group(), aws_user_remove_from_group()

attribute: Arn (includes AWS Account ID)

aws_user_access_key_delete(): Access Key ID