List policies

List policies

Usage

aws_policies(refresh = FALSE, ...)

Arguments

refresh

(logical) refresh results? default: FALSE. to invalidate cache and refresh policy data, set refresh=TRUE

...

named arguments passed on to list_policies

Value

A tibble with information about policies. Each row is a policy. Columns:

PolicyName

PolicyId

Path

Arn

CreateDate

UpdateDate

AttachmentCount

PermissionsBoundaryUsageCount

IsAttachable

Description

Tags

Details

uses memoise internally to cache results to speed up all subsequent calls to the function

See also

Examples

if (FALSE) { # \dontrun{
aws_policies()
aws_policies()
aws_policies(refresh = TRUE)
} # }