Create a user

Create a user

Usage

aws_user_create(username, path = NULL, permission_boundary = NULL, tags = NULL)

Arguments

username

(character) A user name. required

path

(character) The path for the user name. optional. If it is not included, it defaults to a slash (/).

permission_boundary

(character) The ARN of the managed policy that is used to set the permissions boundary for the user. optional

tags

(list) A list of tags that you want to attach to the new user. optional

Value

A tibble with information about the user created

Details

See https://www.paws-r-sdk.com/docs/iam_create_user/ docs for details on the parameters

See also

Examples

if (FALSE) { # \dontrun{
aws_user_create("testBlueBird")
} # }