Create a user

Create a user

Usage

six_user_create(username, path = NULL, permission_boundary = NULL, tags = NULL)

Arguments

username

(character) A user name. required

path

(character) The path for the user name. optional. If it is not included, it defaults to a slash (/).

permission_boundary

(character) The ARN of the managed policy that is used to set the permissions boundary for the user. optional

tags

(list) A list of tags that you want to attach to the new user. optional

Value

NULL invisibly. A draft email is copied to your clipboard

Details

See aws_user_create() for more details. This function creates a user, adds policies so the user can access their own account, and grants them an access key. Add more policies using aws_polic* functions

What is magical

Adds a GetUser policy to your account if doesn't exist yet

Attaches GetUser policy to the user created

Grants an access key

See also

Examples

if (FALSE) { # interactive()
name <- random_user()
six_user_create(name)
}